+++ test/functional/basic/find-e2e-specs.js
@@ -3,12 +3,11 @@ import _ from 'lodash';
 import {retryInterval} from 'asyncbox';
 import {
   extractCapabilityValue,
-  amendCapabilities,
   UICATALOG_CAPS,
   PLATFORM_VERSION,
 } from '../desired';
 import {PREDICATE_SEARCH, CLASS_CHAIN_SEARCH} from '../helpers/element';
-import {initSession, deleteSession, hasDefaultPrebuiltWDA, MOCHA_TIMEOUT} from '../helpers/session';
+import {initSession, deleteSession, MOCHA_TIMEOUT} from '../helpers/session';
 import {util} from 'appium/support';
 
 
@@ -34,10 +33,7 @@ describe('XCUITestDriver - find -', function () {
     chai.should();
     chai.use(chaiAsPromised.default);
 
-    const caps = amendCapabilities(UICATALOG_CAPS, {
-      'appium:usePrebuiltWDA': hasDefaultPrebuiltWDA(),
-    });
-    driver = await initSession(caps);
+    driver = await initSession(UICATALOG_CAPS);
   });
   after(async function () {
     await deleteSession();
@@ -170,7 +166,7 @@ describe('XCUITestDriver - find -', function () {
       });
       it('should return a single element', async function () {
         let el = await driver.$('//XCUIElementTypeButton');
-        (await el.getAttribute('name')).should.equal(APP_TITLE);
+        (await el.getAttribute('label')).should.equal(APP_TITLE);
       });
       it('should return multiple elements', async function () {
         let els = await driver.$$('//XCUIElementTypeButton');