+++ test/functional/helpers/session.js
@@ -1,34 +1,19 @@
-import {extractCapabilityValue} from '../desired';
-
 const HOST = process.env.APPIUM_TEST_SERVER_HOST || '127.0.0.1';
 const PORT = parseInt(String(process.env.APPIUM_TEST_SERVER_PORT), 10) || 4567;
-// on CI the timeout needs to be long, mostly so WDA can be built the first time
-const MOCHA_TIMEOUT = 60 * 1000 * (process.env.CI ? 16 : 4);
+const MOCHA_TIMEOUT = 60 * 1000 * 4;
 
 let driver;
-let didBuildWda = false;
 
 async function initSession(caps, remoteOpts = {}) {
-  try {
-    const {remote} = await import('webdriverio');
-    driver = await remote({
-      hostname: HOST,
-      port: PORT,
-      capabilities: caps,
-      connectionRetryTimeout: MOCHA_TIMEOUT,
-      connectionRetryCount: 1,
-      ...remoteOpts,
-    });
-    if (
-      !extractCapabilityValue(caps, 'appium:usePrebuiltWDA') &&
-      !extractCapabilityValue(caps, 'appium:derivedDataPath')
-    ) {
-      didBuildWda = true;
-    }
-  } catch (e) {
-    didBuildWda = false;
-    throw e;
-  }
+  const {remote} = await import('webdriverio');
+  driver = await remote({
+    hostname: HOST,
+    port: PORT,
+    capabilities: caps,
+    connectionRetryTimeout: MOCHA_TIMEOUT,
+    connectionRetryCount: 1,
+    ...remoteOpts,
+  });
   // @ts-expect-error private API, apparently
   driver.name = undefined;
   // @ts-expect-error private API, apparently
@@ -45,8 +30,4 @@ async function deleteSession() {
   }
 }
 
-function hasDefaultPrebuiltWDA() {
-  return didBuildWda;
-}
-
-export {initSession, deleteSession, hasDefaultPrebuiltWDA, HOST, PORT, MOCHA_TIMEOUT};
+export {initSession, deleteSession, HOST, PORT, MOCHA_TIMEOUT};