+++ test/functional/web/safari-alerts-e2e-specs.js
@@ -1,6 +1,10 @@
 import {retryInterval} from 'asyncbox';
-import {SAFARI_CAPS, amendCapabilities} from '../desired';
-import {initSession, deleteSession, hasDefaultPrebuiltWDA, MOCHA_TIMEOUT} from '../helpers/session';
+import {
+  SAFARI_CAPS,
+  amendCapabilities,
+  isIosVersionBelow
+} from '../desired';
+import {initSession, deleteSession, MOCHA_TIMEOUT} from '../helpers/session';
 import {GUINEA_PIG_PAGE} from './helpers';
 
 
@@ -20,8 +24,12 @@ describe('safari - alerts', function () {
     const caps = amendCapabilities(SAFARI_CAPS, {
       'appium:safariInitialUrl': GUINEA_PIG_PAGE,
       'appium:safariAllowPopups': true,
-      'appium:usePrebuiltWDA': hasDefaultPrebuiltWDA(),
     });
+
+    if (process.env.CI && isIosVersionBelow('18.0')) {
+      this.skip();
+    };
+
     driver = await initSession(caps);
   });
   after(async function () {
@@ -35,8 +43,6 @@ describe('safari - alerts', function () {
   async function dismissAlert(driver) {
     await retryInterval(5, 500, driver.dismissAlert.bind(driver));
   }
-
-  // All tests below are skipped until https://github.com/appium/appium/issues/17013 is resolved
 
   it('should accept alert', async function () {
     const alert = await driver.$('#alert1');